+++ familyhelper-project-sdk/src/main/java/com/dwarfeng/familyhelper/project/sdk/bean/dto/WebInputTaskCreateInfo.java
@@ -9,6 +9,8 @@ import org.checkerframework.checker.index.qual.Positive;
 import org.hibernate.validator.constraints.Length;
 
 import javax.validation.Valid;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
 import java.util.Objects;
 
 /**
@@ -43,6 +45,8 @@ public class WebInputTaskCreateInfo implements Dto {
     private String type;
 
     @JSONField(name = "name")
+    @NotNull
+    @NotEmpty
     @Length(max = Constraints.LENGTH_NAME)
     private String name;