+++ familyhelper-project-sdk/src/main/java/com/dwarfeng/familyhelper/project/sdk/util/Constants.java
@@ -19,6 +19,9 @@ public final class Constants {
     public static final int PERMISSION_LEVEL_GUEST = 1;
     public static final int PERMISSION_LEVEL_MODIFIER = 2;
 
+    public static final int TIME_POINT_STATUS_IN_PROGRESS = 0;
+    public static final int TIME_POINT_STATUS_FINISHED = 1;
+
     private Constants() {
         throw new IllegalStateException("禁止实例化");
     }